Changes in directory structure,API can create user
Changes made in project's directory structure to enable
testing of the API'S. The API can create user and store user
 details.
[Feature #155688213]
:wq
