--- Restfull API/api/version/.pytest_cache/v/cache/lastfailed
@@ -1,3 +1,0 @@
-{
-  "tests/test_auth.py::UserStuffTestCase::test_user_registration": true
-}
